+++ src/entities/mats-data-submission.entity.ts
@@ -35,7 +35,7 @@ export class MatsDataSubmission extends BaseEntity {
 
   @Column({
     name: 'fac_id',
-    type: 'numeric',
+    type: 'bigint',
     transformer: new NumericColumnTransformer(),
   })
   facId: number;
